renaming ui_helpers to u_helpers for UIC
