+++ src/forms/settings/settings.cpp
@@ -14,7 +14,7 @@
 #include "helpers/http_status.h"
 #include "helpers/netman.h"
 #include "helpers/stopreply.h"
-#include "helpers/ui_helpers.h"
+#include "helpers/u_helpers.h"
 #include "hotkeymanager.h"
 #include "components/custom_keyseq_edit/singlestrokekeysequenceedit.h"
